backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
stripe>=10.0
django-import-export>=3.0
requests>=2.20
